requirements.txt
python-consul
hsettings
click
colorama
diskcache
